--- src/main/java/roomescape/infra/payment/dto/PaymentErrorResponse.java
@@ -1,4 +1,0 @@
-package roomescape.infra.payment.dto;
-
-public record PaymentErrorResponse(String code, String message) {
-}
