--- src/main/java/roomescape/infra/payment/exception/PaymentConfirmServerException.java
@@ -1,10 +1,0 @@
-package roomescape.infra.payment.exception;
-
-import roomescape.exception.ApplicationException;
-
-public class PaymentConfirmServerException extends ApplicationException {
-
-    public PaymentConfirmServerException(String message) {
-        super(message);
-    }
-}
