+++ src/main/java/roomescape/infra/payment/toss/dto/TossPaymentErrorResponse.java
@@ -1,0 +1,4 @@
+package roomescape.infra.payment.toss.dto;
+
+public record TossPaymentErrorResponse(String code, String message) {
+}
