+++ src/main/java/roomescape/infra/payment/toss/exception/TossPaymentConfirmException.java
@@ -1,13 +1,13 @@
-package roomescape.infra.payment.exception;
+package roomescape.infra.payment.toss.exception;
 
 import org.springframework.http.HttpStatusCode;
 import roomescape.exception.ApplicationException;
 
-public class PaymentConfirmException extends ApplicationException {
+public class TossPaymentConfirmException extends ApplicationException {
 
     private final HttpStatusCode statusCode;
 
-    public PaymentConfirmException(String message, HttpStatusCode statusCode) {
+    public TossPaymentConfirmException(String message, HttpStatusCode statusCode) {
         super(message);
         this.statusCode = statusCode;
     }
